Added lots of functionality to DataHandler.js
We added 2 new arrays, formattedData which has the data from the csv file formatted with the right types and sortedDateByDate which has the whole array sorted by dates. We also added getEmailsByDate which returns emails between a specific timeframe. Lastly we added getFirstEmailDate, getLastEmailDate, and getEmailDateByPercentile which return the first email, the last email, and an email on a specific percentile from the data array.

Co-Authored-By: Basmeuw <32522618+Basmeuw@users.noreply.github.com>