Added a delay for drawing the graph
For some reason does the node link diagram not automatically draw itself correctly, but after 5 seconds it can be drawn correctly. So i added a wait when the graph is first drawn which draws the graph correctly after 5 seconds. This is ofcourse not a finite solution but it works for now.