Fixed selection when grouped on jobtitle in nodelink diagram and fixed tooltip when grouped on jobtitle in nodelink diagram
The selection was not working when the nodelink diagram is grouped on jobtitle. This is now fixed. The selection works now across all visualizations except when you select the big node. Then only the edges in the nodelink diagram are highlighted. Also the tooltip broke when the node was a grouped one. this is also fixed now.