Highlighting also works with big nodes now. Also fixed problem with data not changing
When you click on an big node in the nodelink diagram, all persons grouped in that big node get added to datahandler.highlightpersons, which also is changed from an int to an array. Fixed that all visualizations work with an array of highlighted persons. I also fixed a problem i created in the previous pushes that the visualizations did not always update.